refactor(mobile): change useNonBlacklistedDistribution from general.ts to a file in hooks folder
