+++ apps/ledger-live-mobile/src/actions/general.ts
@@ -159,24 +159,3 @@ export function useTrackingPairs(): TrackingPair[] {
     [extraSessionTrackingPairs, trPairs],
   );
 }
-
-export function useNonBlacklistedDistribution(
-  opts: Omit<Parameters<typeof useDistributionCommon>[0], "accounts" | "to"> = {
-    showEmptyAccounts: true,
-  },
-) {
-  const distribution = useDistribution(opts);
-  const blacklistedTokenIds = useSelector(blacklistedTokenIdsSelector);
-  const blacklistedTokenIdsSet = useMemo(() => new Set(blacklistedTokenIds), [blacklistedTokenIds]);
-
-  return useMemo(
-    () =>
-      distribution.isAvailable
-        ? distribution.list.filter(
-            ({ currency }) =>
-              currency.type !== "TokenCurrency" || !blacklistedTokenIdsSet.has(currency.id),
-          )
-        : [],
-    [distribution.isAvailable, distribution.list, blacklistedTokenIdsSet],
-  );
-}
