+++ apps/ledger-live-mobile/src/mvvm/features/Analytics/__integrations__/detailedAllocation.integration.test.tsx
@@ -7,9 +7,23 @@ import {
   mockBitcoinCurrency,
   mockEthereumCurrency,
   mockCardanoCurrency,
-  mockEthAccountWithUSDC,
   usdcToken,
 } from "./shared";
+import {
+  createFixtureAccount,
+  createFixtureTokenAccount,
+} from "@ledgerhq/live-common/mock/fixtures/cryptoCurrencies";
+import type { Account } from "@ledgerhq/types-live";
+
+const createMockEthAccountWithUSDC = (): Account => {
+  const ethAccount = createFixtureAccount("01");
+  const usdcSubAccount = createFixtureTokenAccount("01", usdcToken);
+  return {
+    ...ethAccount,
+    subAccounts: [{ ...usdcSubAccount, parentId: ethAccount.id }],
+  };
+};
+
 import { track } from "~/analytics";
 import { NavigatorName, ScreenName } from "~/const";
 
@@ -112,16 +126,26 @@ describe("DetailedAllocation Integration Tests", () => {
   });
 
   describe("blacklisted tokens", () => {
-    const mockStateWithBlacklistedToken = (state: State): State => ({
-      ...state,
-      accounts: { ...state.accounts, active: [mockEthAccountWithUSDC] },
-      settings: { ...state.settings, counterValue: "USD", blacklistedTokenIds: [usdcToken.id] },
-    });
 
-    const mockStateWithNonBlacklistedToken = (state: State): State => ({
-      ...state,
-      accounts: { ...state.accounts, active: [mockEthAccountWithUSDC] },
-      settings: { ...state.settings, counterValue: "USD", blacklistedTokenIds: [] },
+    const mockStateWithBlacklistedToken = (state: State): State => {
+     const mockEthAccountWithUSDC = createMockEthAccountWithUSDC();
+      return {
+        ...state,
+        accounts: { ...state.accounts, active: [mockEthAccountWithUSDC] },
+        settings: { ...state.settings, counterValue: "USD", blacklistedTokenIds: [usdcToken.id] },
+      };
+    }
+    const mockStateWithNonBlacklistedToken = (state: State): State => {
+     const mockEthAccountWithUSDC = createMockEthAccountWithUSDC();
+      return {
+        ...state,
+        accounts: { ...state.accounts, active: [mockEthAccountWithUSDC] },
+        settings: { ...state.settings, counterValue: "USD", blacklistedTokenIds: [] },
+      };
+    }
+
+    beforeEach(() => {
+      jest.clearAllMocks();
     });
 
     it("should display a token that is not blacklisted", () => {
