+++ apps/ledger-live-mobile/src/mvvm/features/Analytics/__integrations__/shared.tsx
@@ -3,7 +3,6 @@ import Navigator from "../Navigator";
 import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets";
 import { genAccount } from "@ledgerhq/live-common/mock/account";
 import { usdcToken } from "@ledgerhq/live-common/modularDrawer/__mocks__/currencies.mock";
-import { createFixtureEthAccountWithSubToken } from "@ledgerhq/live-common/mock/fixtures/cryptoCurrencies";
 import type { Account } from "@ledgerhq/types-live";
 import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 
@@ -16,8 +15,6 @@ export const TestNavigatorWrapper: React.FC = () => {
 export const mockBitcoinCurrency = getCryptoCurrencyById("bitcoin");
 export const mockEthereumCurrency = getCryptoCurrencyById("ethereum");
 export const mockCardanoCurrency = getCryptoCurrencyById("cardano");
-
-export const mockEthAccountWithUSDC: Account = createFixtureEthAccountWithSubToken(usdcToken);
 
 export const createMockAccount = (currency: CryptoCurrency, id: string): Account => {
   return genAccount(id, { currency, operationsSize: 0 });
