+++ apps/ledger-live-mobile/src/mvvm/features/Assets/hooks/useAssetsListViewModel.ts
@@ -2,7 +2,8 @@ import { useCallback, useMemo } from "react";
 import useEnv from "@ledgerhq/live-common/hooks/useEnv";
 import { useFocusEffect, useNavigation } from "@react-navigation/native";
 import { GestureResponderEvent } from "react-native";
-import { useNonBlacklistedDistribution, useRefreshAccountsOrdering } from "~/actions/general";
+import { useNonBlacklistedDistribution } from "~/hooks/useNonBlacklistedDistribution";
+import { useRefreshAccountsOrdering } from "~/actions/general";
 import { NavigatorName, ScreenName } from "~/const";
 import { Asset } from "~/types/asset";
 import { track } from "~/analytics";
