+++ apps/ledger-live-mobile/src/types/distributionOptions.ts
@@ -1,0 +1,4 @@
+export type DistributionOptions = {
+  showEmptyAccounts?: boolean;
+  hideEmptyTokenAccount?: boolean;
+};
