+++ libs/ledger-live-common/src/mock/fixtures/cryptoCurrencies.ts
@@ -38,7 +38,6 @@ export function createFixtureCryptoCurrency(family: string): CryptoCurrency {
 }
 
 const defaultEthCryptoFamily = cryptocurrenciesById["ethereum"];
-
 const defaultERC20USDTToken: TokenCurrency = {
   type: "TokenCurrency",
   id: "ethereum/erc20/usd_tether__erc20_",
@@ -121,14 +120,5 @@ export function createFixtureAccount(
       },
     },
     swapHistory: [],
-  };
-}
-
-export function createFixtureEthAccountWithSubToken(token: TokenCurrency): Account {
-  const ethAccount = createFixtureAccount("01");
-  const subAccount = createFixtureTokenAccount("01", token);
-  return {
-    ...ethAccount,
-    subAccounts: [{ ...subAccount, parentId: ethAccount.id }],
   };
 }
